Fixing RateLimit.cs
Currently this class is broken, double.TryParsing is not parsing correctly the headers which leads to an unhandled exception to be thrown, this is because it ignores the dot based on the current culture as it wasn't specified to allow them. This issue causes TimeSpan to overflow, because "1582246711.602" gets parsed into 1582246711602 instead of 1582246711.602 making it to add over 50k years instead of 50.

this issue also happens with X-RateLimit-Reset-After to have a wrong value.