Implement application emojis (#123)
* Implemented application emojis.

* Removed unnecessary using statement

Left over from when extensions were going to be on ICurrentUser...

* Fixed FetchApplicationEmojisAsync throwing

GET /applications/{application.id}/emojis returns the emojis under a nested "items" object instead of returning an array directly. oops

* IApplicationEntity should inherit from IEntity

* Abstracted IGuildEmoji and IApplicationEmoji -> IOwnedEmoji

IOwnedEmoji as a name was chosen since it represents an emoji that is (potentially) *owned* by a guild or application. Alternatives considered included IDiscordEmoji, but that seemed way too vague and not very descriptive of its common properties.

Also, adjusted XML documentation to compensate for these changes.